pyproject.toml
[project]
name = "desk-thing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "pyserial>=3.5",
    "winrt-runtime==3.2.1",
    "winrt-windows-foundation==3.2.1",
    "winrt-windows-foundation-collections==3.2.1",
    "winrt-windows-media-control==3.2.1",
    "winrt-windows-media-playback==3.2.1",
    "winrt-windows-storage==3.2.1",
    "winrt-windows-storage-streams==3.2.1",
]
